@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
@@ -27,9 +27,14 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Controllers\" />
-    <Folder Include="Servicos\" />
-    <Folder Include="Repositorio\" />
+    <Folder Include="Interfaces\Empresa\Operacoes\" />
+    <Folder Include="Migrations\" />
+    <Folder Include="Repositorio\Empresa\Operacoes\" />
+    <Folder Include="Servicos\Empresa\Operacoes\" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Common.csproj" />
   </ItemGroup>
 
 </Project>
