@@ -1,6 +1,6 @@
-using Api.DTO.Requisicao.Usuario.Cadastro;
-using Api.DTO.Resposta.Usuario.Cadastro;
 using Api.Interfaces.Usuario.Cadastro;
+using Common.DTO.Requisicao.Usuario.Cadastro;
+using Common.DTO.Resposta.Usuario.Cadastro;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
