@@ -1,14 +1,15 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
-using Api.DTO.Requisicao.Usuario.Login;
-using Api.DTO.Resposta.Usuario.Login;
-using Api.Enums.Usuario;
+using Common.DTO.Requisicao.Usuario.Login;
+using Common.Enums.USUARIO;
 using Api.Interfaces.Usuario.Login;
 using Api.JWT;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
+using Common.DTO.Resposta.Usuario.Login;
+using Api.Interfaces.Usuario.Operacoes;
 
 namespace Api.Controllers
 {
@@ -18,12 +19,14 @@ namespace Api.Controllers
     public class UsuarioLoginController : ControllerBase
     {
         private readonly IUsuarioLoginServico _loginServico;
+        private readonly IUsuarioOperacoesServico _usuarioOperacoes;
         private readonly IConfiguration _configuration;
 
-        public UsuarioLoginController(IUsuarioLoginServico loginServico, IConfiguration configuration)
+        public UsuarioLoginController(IUsuarioLoginServico loginServico, IConfiguration configuration, IUsuarioOperacoesServico usuarioOperacoes)
         {
             _loginServico = loginServico;
             _configuration = configuration;
+            _usuarioOperacoes = usuarioOperacoes;
         }
 
         [AllowAnonymous]
@@ -35,13 +38,15 @@ namespace Api.Controllers
             if(resposta.SUCESSO)
             {
                 var usuarioID = await _loginServico.BuscarUsuario(requisicao.EMAIL, requisicao.SENHA);
+                var usuario = await _usuarioOperacoes.BuscarUsuarioBrutoPorID(usuarioID);
 
                 var token = GerarTokenJWT(
                     secretKey: ChaveJWT.PegarChaveSecreta(_configuration), 
                     userId: usuarioID,
                     issuer: "InovarJuntoAPI",
                     audience: "InovarJuntoFrontend",
-                    expireInMinutes: 120 
+                    expireInMinutes: 120,
+                    permissaoId: usuario.PERMISSAOID
                     );
 
                 resposta.TOKEN = token;
@@ -49,10 +54,10 @@ namespace Api.Controllers
                 return Ok(resposta);
             }
 
-            return BadRequest();
+            return Unauthorized();
         }
 
-        private static string GerarTokenJWT(string secretKey, int expireInMinutes, string issuer, string audience, int userId)
+        private static string GerarTokenJWT(string secretKey, int expireInMinutes, string issuer, string audience, int userId, int permissaoId)
         {
             var tokenHandler = new JwtSecurityTokenHandler();
             var claims = new[]
@@ -60,7 +65,7 @@ namespace Api.Controllers
                 new Claim(JwtRegisteredClaimNames.Sub, userId.ToString()),
                 new Claim(JwtRegisteredClaimNames.Jti, Guid.NewGuid().ToString()),
                 new Claim(JwtRegisteredClaimNames.Iat, ToUnixEpochDate(DateTime.UtcNow).ToString(), ClaimValueTypes.Integer64),
-                new Claim(ClaimTypes.Role, EPERMISSAO.USUARIO.ToString())
+                new Claim("roleId", permissaoId.ToString())
             };
 
             var tokenDecriptor = new SecurityTokenDescriptor
