@@ -1,5 +1,5 @@
-﻿using Api.DTO.Requisicao.Usuario.Cadastro;
-using Api.DTO.Resposta.Usuario.Cadastro;
+﻿using Common.DTO.Requisicao.Usuario.Cadastro;
+using Common.DTO.Resposta.Usuario.Cadastro;
 
 namespace Api.Interfaces.Usuario.Cadastro
 {
