@@ -1,6 +1,6 @@
-﻿using Api.DTO.Requisicao.Usuario.Login;
-using Api.DTO.Resposta.Usuario.Login;
-using Api.Models;
+﻿using Common.DTO.Requisicao.Usuario.Login;
+using Common.Models;
+using Common.DTO.Resposta.Usuario.Login;
 
 namespace Api.Interfaces.Usuario.Login
 {
