@@ -17,12 +17,12 @@ namespace Api.Migrations
         {
 #pragma warning disable 612, 618
             modelBuilder
-                .HasAnnotation("ProductVersion", "8.0.8")
+                .HasAnnotation("ProductVersion", "8.0.10")
                 .HasAnnotation("Relational:MaxIdentifierLength", 64);
 
             MySqlModelBuilderExtensions.AutoIncrementColumns(modelBuilder);
 
-            modelBuilder.Entity("Api.Models.CATEGORIA", b =>
+            modelBuilder.Entity("Common.Models.CATEGORIA", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -38,7 +38,7 @@ namespace Api.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("varchar(255)");
 
-                    b.Property<int>("EMPRESA_ID")
+                    b.Property<int>("EMPRESAID")
                         .HasColumnType("int");
 
                     b.Property<string>("NOME")
@@ -46,19 +46,19 @@ namespace Api.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("varchar(100)");
 
-                    b.Property<int>("USUARIO_ID")
+                    b.Property<int>("USUARIOID")
                         .HasColumnType("int");
 
                     b.HasKey("ID");
 
-                    b.HasIndex("EMPRESA_ID");
+                    b.HasIndex("EMPRESAID");
 
-                    b.HasIndex("USUARIO_ID");
+                    b.HasIndex("USUARIOID");
 
                     b.ToTable("CATEGORIA");
                 });
 
-            modelBuilder.Entity("Api.Models.EMPRESA", b =>
+            modelBuilder.Entity("Common.Models.EMPRESA", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -68,8 +68,7 @@ namespace Api.Migrations
 
                     b.Property<string>("CNPJ")
                         .IsRequired()
-                        .HasMaxLength(15)
-                        .HasColumnType("varchar(15)");
+                        .HasColumnType("longtext");
 
                     b.Property<DateTime>("DATA_CRIACAO_CADASTRO")
                         .HasColumnType("datetime(6)");
@@ -79,24 +78,21 @@ namespace Api.Migrations
 
                     b.Property<string>("EMAIL")
                         .IsRequired()
-                        .HasMaxLength(50)
-                        .HasColumnType("varchar(50)");
+                        .HasColumnType("longtext");
 
                     b.Property<int>("EMPRESA_ATIVA")
                         .HasColumnType("int");
 
-                    b.Property<int>("ENDERECO_ID")
+                    b.Property<int>("ENDERECOID")
                         .HasColumnType("int");
 
                     b.Property<string>("NOME_FANTASIA")
                         .IsRequired()
-                        .HasMaxLength(255)
-                        .HasColumnType("varchar(255)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("RAZAO_SOCIAL")
                         .IsRequired()
-                        .HasMaxLength(255)
-                        .HasColumnType("varchar(255)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("TELEFONE")
                         .IsRequired()
@@ -104,12 +100,12 @@ namespace Api.Migrations
 
                     b.HasKey("ID");
 
-                    b.HasIndex("ENDERECO_ID");
+                    b.HasIndex("ENDERECOID");
 
                     b.ToTable("EMPRESA");
                 });
 
-            modelBuilder.Entity("Api.Models.ENDERECO", b =>
+            modelBuilder.Entity("Common.Models.ENDERECO", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -119,43 +115,37 @@ namespace Api.Migrations
 
                     b.Property<string>("BAIRRO")
                         .IsRequired()
-                        .HasMaxLength(50)
-                        .HasColumnType("varchar(50)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("CEP")
                         .IsRequired()
-                        .HasMaxLength(8)
-                        .HasColumnType("varchar(8)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("CIDADE")
                         .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("varchar(100)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("ESTADO")
                         .IsRequired()
-                        .HasMaxLength(20)
-                        .HasColumnType("varchar(20)");
+                        .HasColumnType("longtext");
 
                     b.Property<int>("NUMERO")
                         .HasColumnType("int");
 
                     b.Property<string>("PAIS")
                         .IsRequired()
-                        .HasMaxLength(30)
-                        .HasColumnType("varchar(30)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("RUA")
                         .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("varchar(100)");
+                        .HasColumnType("longtext");
 
                     b.HasKey("ID");
 
                     b.ToTable("ENDERECO");
                 });
 
-            modelBuilder.Entity("Api.Models.FUNCIONARIO", b =>
+            modelBuilder.Entity("Common.Models.FUNCIONARIO", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -187,7 +177,7 @@ namespace Api.Migrations
                         .HasMaxLength(50)
                         .HasColumnType("varchar(50)");
 
-                    b.Property<int>("ENDERECO_ID")
+                    b.Property<int>("ENDERECOID")
                         .HasColumnType("int");
 
                     b.Property<string>("FOTO")
@@ -204,10 +194,10 @@ namespace Api.Migrations
                         .HasMaxLength(100)
                         .HasColumnType("varchar(100)");
 
-                    b.Property<int>("PERMISSAO_ID")
+                    b.Property<int>("PERMISSAOID")
                         .HasColumnType("int");
 
-                    b.Property<int>("SENHA_ID")
+                    b.Property<int>("SENHAID")
                         .HasColumnType("int");
 
                     b.Property<string>("SOBRENOME")
@@ -227,16 +217,16 @@ namespace Api.Migrations
                     b.HasIndex("EMAIL")
                         .IsUnique();
 
-                    b.HasIndex("ENDERECO_ID");
+                    b.HasIndex("ENDERECOID");
 
-                    b.HasIndex("PERMISSAO_ID");
+                    b.HasIndex("PERMISSAOID");
 
-                    b.HasIndex("SENHA_ID");
+                    b.HasIndex("SENHAID");
 
                     b.ToTable("FUNCIONARIO");
                 });
 
-            modelBuilder.Entity("Api.Models.HISTORICO_MOVIMENTACOES", b =>
+            modelBuilder.Entity("Common.Models.HISTORICO_MOVIMENTACOES", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -252,10 +242,10 @@ namespace Api.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("varchar(255)");
 
-                    b.Property<int>("FUNCIONARIO_ID")
+                    b.Property<int>("FUNCIONARIOID")
                         .HasColumnType("int");
 
-                    b.Property<int>("PRODUTO_ID")
+                    b.Property<int>("PRODUTOID")
                         .HasColumnType("int");
 
                     b.Property<int>("QUANTIDADE")
@@ -264,21 +254,21 @@ namespace Api.Migrations
                     b.Property<int>("TIPO_MOVIMENTACAO")
                         .HasColumnType("int");
 
-                    b.Property<int>("USUARIO_ID")
+                    b.Property<int>("USUARIOID")
                         .HasColumnType("int");
 
                     b.HasKey("ID");
 
-                    b.HasIndex("FUNCIONARIO_ID");
+                    b.HasIndex("FUNCIONARIOID");
 
-                    b.HasIndex("PRODUTO_ID");
+                    b.HasIndex("PRODUTOID");
 
-                    b.HasIndex("USUARIO_ID");
+                    b.HasIndex("USUARIOID");
 
                     b.ToTable("HISTORICO_MOVIMENTACOES");
                 });
 
-            modelBuilder.Entity("Api.Models.PERMISSAO", b =>
+            modelBuilder.Entity("Common.Models.PERMISSAO", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -327,7 +317,7 @@ namespace Api.Migrations
                         });
                 });
 
-            modelBuilder.Entity("Api.Models.PRODUTO", b =>
+            modelBuilder.Entity("Common.Models.PRODUTO", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -338,7 +328,7 @@ namespace Api.Migrations
                     b.Property<int>("ATIVO")
                         .HasColumnType("int");
 
-                    b.Property<int>("CATEGORIA_ID")
+                    b.Property<int>("CATEGORIAID")
                         .HasColumnType("int");
 
                     b.Property<DateTime>("DATA_CRIACAO")
@@ -349,7 +339,7 @@ namespace Api.Migrations
                         .HasMaxLength(255)
                         .HasColumnType("varchar(255)");
 
-                    b.Property<int>("EMPRESA_ID")
+                    b.Property<int>("EMPRESAID")
                         .HasColumnType("int");
 
                     b.Property<string>("NOME")
@@ -360,21 +350,21 @@ namespace Api.Migrations
                     b.Property<decimal>("PRECO")
                         .HasColumnType("decimal(65,30)");
 
-                    b.Property<int>("USUARIO_ID")
+                    b.Property<int>("USUARIOID")
                         .HasColumnType("int");
 
                     b.HasKey("ID");
 
-                    b.HasIndex("CATEGORIA_ID");
+                    b.HasIndex("CATEGORIAID");
 
-                    b.HasIndex("EMPRESA_ID");
+                    b.HasIndex("EMPRESAID");
 
-                    b.HasIndex("USUARIO_ID");
+                    b.HasIndex("USUARIOID");
 
                     b.ToTable("PRODUTO");
                 });
 
-            modelBuilder.Entity("Api.Models.SENHA", b =>
+            modelBuilder.Entity("Common.Models.SENHA", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -399,7 +389,7 @@ namespace Api.Migrations
                     b.ToTable("SENHA");
                 });
 
-            modelBuilder.Entity("Api.Models.USUARIO", b =>
+            modelBuilder.Entity("Common.Models.USUARIO", b =>
                 {
                     b.Property<int>("ID")
                         .ValueGeneratedOnAdd()
@@ -409,50 +399,42 @@ namespace Api.Migrations
 
                     b.Property<string>("CPF")
                         .IsRequired()
-                        .HasMaxLength(11)
-                        .HasColumnType("varchar(11)");
+                        .HasColumnType("longtext");
 
                     b.Property<DateTime>("DATA_CRIADO")
                         .HasColumnType("datetime(6)");
 
                     b.Property<string>("EMAIL")
                         .IsRequired()
-                        .HasMaxLength(50)
-                        .HasColumnType("varchar(50)");
+                        .HasColumnType("varchar(255)");
 
-                    b.Property<int>("ENDERECO_ID")
+                    b.Property<int>("ENDERECOID")
                         .HasColumnType("int");
 
                     b.Property<string>("FOTO")
-                        .IsRequired()
                         .HasColumnType("longtext");
 
-                    b.Property<int?>("FUNCIONARIO_ID")
+                    b.Property<int?>("FUNCIONARIOID")
                         .HasColumnType("int");
 
                     b.Property<string>("NOME")
                         .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("varchar(100)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("NOME_SOCIAL")
-                        .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("varchar(100)");
+                        .HasColumnType("longtext");
 
-                    b.Property<int>("PERMISSAO_ID")
+                    b.Property<int>("PERMISSAOID")
                         .HasColumnType("int");
 
-                    b.Property<int>("SENHA_ID")
+                    b.Property<int>("SENHAID")
                         .HasColumnType("int");
 
                     b.Property<string>("SOBRENOME")
                         .IsRequired()
-                        .HasMaxLength(100)
-                        .HasColumnType("varchar(100)");
+                        .HasColumnType("longtext");
 
                     b.Property<string>("TELEFONE")
-                        .IsRequired()
                         .HasColumnType("longtext");
 
                     b.Property<bool>("USUARIO_ATIVO")
@@ -463,28 +445,28 @@ namespace Api.Migrations
                     b.HasIndex("EMAIL")
                         .IsUnique();
 
-                    b.HasIndex("ENDERECO_ID");
+                    b.HasIndex("ENDERECOID");
 
-                    b.HasIndex("FUNCIONARIO_ID");
+                    b.HasIndex("FUNCIONARIOID");
 
-                    b.HasIndex("PERMISSAO_ID");
+                    b.HasIndex("PERMISSAOID");
 
-                    b.HasIndex("SENHA_ID");
+                    b.HasIndex("SENHAID");
 
                     b.ToTable("USUARIO");
                 });
 
-            modelBuilder.Entity("Api.Models.CATEGORIA", b =>
+            modelBuilder.Entity("Common.Models.CATEGORIA", b =>
                 {
-                    b.HasOne("Api.Models.EMPRESA", "EMPRESA")
+                    b.HasOne("Common.Models.EMPRESA", "EMPRESA")
                         .WithMany()
-                        .HasForeignKey("EMPRESA_ID")
+                        .HasForeignKey("EMPRESAID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.USUARIO", "USUARIO")
+                    b.HasOne("Common.Models.USUARIO", "USUARIO")
                         .WithMany()
-                        .HasForeignKey("USUARIO_ID")
+                        .HasForeignKey("USUARIOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
@@ -493,34 +475,34 @@ namespace Api.Migrations
                     b.Navigation("USUARIO");
                 });
 
-            modelBuilder.Entity("Api.Models.EMPRESA", b =>
+            modelBuilder.Entity("Common.Models.EMPRESA", b =>
                 {
-                    b.HasOne("Api.Models.ENDERECO", "ENDERECO")
+                    b.HasOne("Common.Models.ENDERECO", "ENDERECO")
                         .WithMany()
-                        .HasForeignKey("ENDERECO_ID")
+                        .HasForeignKey("ENDERECOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
                     b.Navigation("ENDERECO");
                 });
 
-            modelBuilder.Entity("Api.Models.FUNCIONARIO", b =>
+            modelBuilder.Entity("Common.Models.FUNCIONARIO", b =>
                 {
-                    b.HasOne("Api.Models.ENDERECO", "ENDERECO")
+                    b.HasOne("Common.Models.ENDERECO", "ENDERECO")
                         .WithMany()
-                        .HasForeignKey("ENDERECO_ID")
+                        .HasForeignKey("ENDERECOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.PERMISSAO", "PERMISSAO")
+                    b.HasOne("Common.Models.PERMISSAO", "PERMISSAO")
                         .WithMany()
-                        .HasForeignKey("PERMISSAO_ID")
+                        .HasForeignKey("PERMISSAOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.SENHA", "SENHA")
+                    b.HasOne("Common.Models.SENHA", "SENHA")
                         .WithMany()
-                        .HasForeignKey("SENHA_ID")
+                        .HasForeignKey("SENHAID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
@@ -531,23 +513,23 @@ namespace Api.Migrations
                     b.Navigation("SENHA");
                 });
 
-            modelBuilder.Entity("Api.Models.HISTORICO_MOVIMENTACOES", b =>
+            modelBuilder.Entity("Common.Models.HISTORICO_MOVIMENTACOES", b =>
                 {
-                    b.HasOne("Api.Models.FUNCIONARIO", "FUNCIONARIO")
+                    b.HasOne("Common.Models.FUNCIONARIO", "FUNCIONARIO")
                         .WithMany()
-                        .HasForeignKey("FUNCIONARIO_ID")
+                        .HasForeignKey("FUNCIONARIOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.PRODUTO", "PRODUTO")
+                    b.HasOne("Common.Models.PRODUTO", "PRODUTO")
                         .WithMany()
-                        .HasForeignKey("PRODUTO_ID")
+                        .HasForeignKey("PRODUTOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.USUARIO", "USUARIO")
+                    b.HasOne("Common.Models.USUARIO", "USUARIO")
                         .WithMany()
-                        .HasForeignKey("USUARIO_ID")
+                        .HasForeignKey("USUARIOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
@@ -558,23 +540,23 @@ namespace Api.Migrations
                     b.Navigation("USUARIO");
                 });
 
-            modelBuilder.Entity("Api.Models.PRODUTO", b =>
+            modelBuilder.Entity("Common.Models.PRODUTO", b =>
                 {
-                    b.HasOne("Api.Models.CATEGORIA", "CATEGORIA")
+                    b.HasOne("Common.Models.CATEGORIA", "CATEGORIA")
                         .WithMany()
-                        .HasForeignKey("CATEGORIA_ID")
+                        .HasForeignKey("CATEGORIAID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.EMPRESA", "EMPRESA")
+                    b.HasOne("Common.Models.EMPRESA", "EMPRESA")
                         .WithMany()
-                        .HasForeignKey("EMPRESA_ID")
+                        .HasForeignKey("EMPRESAID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.USUARIO", "USUARIO")
+                    b.HasOne("Common.Models.USUARIO", "USUARIO")
                         .WithMany()
-                        .HasForeignKey("USUARIO_ID")
+                        .HasForeignKey("USUARIOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
@@ -585,27 +567,27 @@ namespace Api.Migrations
                     b.Navigation("USUARIO");
                 });
 
-            modelBuilder.Entity("Api.Models.USUARIO", b =>
+            modelBuilder.Entity("Common.Models.USUARIO", b =>
                 {
-                    b.HasOne("Api.Models.ENDERECO", "ENDERECO")
+                    b.HasOne("Common.Models.ENDERECO", "ENDERECO")
                         .WithMany()
-                        .HasForeignKey("ENDERECO_ID")
+                        .HasForeignKey("ENDERECOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.FUNCIONARIO", "FUNCIONARIO")
+                    b.HasOne("Common.Models.FUNCIONARIO", "FUNCIONARIO")
                         .WithMany()
-                        .HasForeignKey("FUNCIONARIO_ID");
+                        .HasForeignKey("FUNCIONARIOID");
 
-                    b.HasOne("Api.Models.PERMISSAO", "PERMISSAO")
+                    b.HasOne("Common.Models.PERMISSAO", "PERMISSAO")
                         .WithMany()
-                        .HasForeignKey("PERMISSAO_ID")
+                        .HasForeignKey("PERMISSAOID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
-                    b.HasOne("Api.Models.SENHA", "SENHA")
+                    b.HasOne("Common.Models.SENHA", "SENHA")
                         .WithMany()
-                        .HasForeignKey("SENHA_ID")
+                        .HasForeignKey("SENHAID")
                         .OnDelete(DeleteBehavior.Cascade)
                         .IsRequired();
 
