@@ -1,10 +1,10 @@
 ﻿using Api.Dados;
-using Api.Models;
-using Api.DTO.Requisicao.Usuario.Login;
-using Api.DTO.Resposta.Usuario.Login;
+using Common.Models;
+using Common.DTO.Requisicao.Usuario.Login;
 using Api.Interfaces.Usuario.Login;
 using Microsoft.EntityFrameworkCore;
-using Api.Seguranca;
+using Common.Seguranca;
+using Common.DTO.Resposta.Usuario.Login;
 
 namespace Api.Repositorio.Usuario.Login
 {
