@@ -1,6 +1,6 @@
-﻿using Api.DTO.Requisicao.Usuario.Operacoes;
+﻿using Common.DTO.Requisicao.Usuario.Operacoes;
 using Api.Interfaces.Usuario.Operacoes;
-using Api.Models;
+using Common.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Api.Servicos.Usuario.Operacoes
