@@ -1,13 +1,23 @@
-﻿<!DOCTYPE html>
+﻿@using Common.Enums
+@inject WebApp.Interfaces.ITokenService TokenService 
+
+@{
+    var token = TokenService.PegarToken();
+    var dados = WebApp.JWT.TokenConfig.DecodificarToken(token);
+}
+
+<!DOCTYPE html>
 <html lang="pt-br">
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>@ViewData["Title"] - InovarJunto</title>
     <script type="importmap"></script>
+    <link rel="stylesheet" href="~/css/global.css" asp-append-version="true" />
     <link rel="stylesheet" href="~/lib/bootstrap/dist/css/zephyr/bootstrap.min.css" />
     <link rel="stylesheet" href="~/css/site.css" asp-append-version="true" />
     <link rel="stylesheet" href="~/WebApp.styles.css" asp-append-version="true" />
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
 </head>
 <body class="bg-light">
     <header>
@@ -21,13 +31,24 @@
                 <div class="navbar-collapse collapse d-sm-inline-flex justify-content-between">
                     <ul class="navbar-nav flex-grow-1">
                         <li class="nav-item">
-                            <a class="nav-link" asp-area="" asp-page="/Index">Principal</a>
+                            <a class="nav-link" asp-area="" asp-page="/Principal">Principal</a>
                         </li>
+                        @if(dados.Permissao == "1")
+                        {
+                            <li class="nav-item">
+                                <a class="nav-link" asp-area="" asp-page="/Admin/Administrativo">Administrativo</a>
+                            </li>
+                        }
                         <li class="nav-item">
                             <a class="nav-link" asp-area="" asp-page="/Privacy">Privacidade</a>
                         </li>
                     </ul>
                 </div>
+                <form asp-page="/Principal" method="post" asp-route-returnUrl="@Url.Page("/Account/Login", new { area = "" })">
+                    <div>
+                        <button class="btn btn-danger">Logout</button>
+                    </div>
+                </form>
             </div>
         </nav>
     </header>
@@ -37,9 +58,9 @@
         </main>
     </div>
 
-    <footer class="border-top footer text-muted">
+    <footer class="footer text-muted bg-dark">
         <div class="container">
-            &copy; @DateTime.Now.Year - InovarJunto - <a asp-area="" asp-page="/Privacy">Privacidade</a>
+            <span class="text-white">&copy; @DateTime.Now.Year - InovarJunto - <a asp-area="" asp-page="/Privacy">Privacidade</a></span>
         </div>
     </footer>
 
