@@ -1,22 +1,55 @@
-var builder = WebApplication.CreateBuilder(args);
+using WebApp.Interfaces;
+using WebApp.JWT;
+using WebApp.Pages;
+using WebApp.Services;
 
-// Add services to the container.
-builder.Services.AddRazorPages();
+var builder = WebApplication.CreateBuilder(args);
+var services = builder.Services;
+var configuration = builder.Configuration;
+
+services.AddHttpClient("extensionistaAPI", client =>
+{
+    client.BaseAddress = new Uri("https://localhost:7071/api/");
+    client.DefaultRequestHeaders.Add("Accept", "application/json");
+});
+
+services.AddSession(opt =>
+{
+    opt.IdleTimeout = TimeSpan.FromMinutes(30);
+    opt.Cookie.Name = "AspNetCore.Session";
+    opt.Cookie.HttpOnly = true;
+    opt.Cookie.IsEssential = true;
+    opt.Cookie.SecurePolicy = CookieSecurePolicy.Always;
+});
+
+services.AddHttpContextAccessor();
+services.AddScoped<ITokenService, TokenService>();
+services.AddTransient<JwtTokenHandler>();
+
+services.AddHttpClient<PrincipalModel>()
+    .AddHttpMessageHandler<JwtTokenHandler>();
+
+services.AddHttpClient();
+services.AddRazorPages(opt =>
+{
+    opt.Conventions.AddPageRoute("/Account/Login", "");
+});
+
 
 var app = builder.Build();
 
-// Configure the HTTP request pipeline.
 if (!app.Environment.IsDevelopment())
 {
     app.UseExceptionHandler("/Error");
-    // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
 
 app.UseHttpsRedirection();
 
 app.UseRouting();
+app.UseSession();
 
+app.UseAuthentication();
 app.UseAuthorization();
 
 app.MapStaticAssets();
